basic tests

See merge request xkopeck5/timesheet-for-freelancer!14

+++ src/main/java/pv168/freelancer/data/WorkDao.java
@@ -1,6 +1,5 @@
 package pv168.freelancer.data;
 
-
 import pv168.freelancer.model.WorkDone;
 import pv168.freelancer.model.WorkType;
 
@@ -20,18 +19,24 @@ public class WorkDao {
     }
 
     public void create(WorkDone workDone) {
+        if (workDone.getId() != null) {
+            throw new IllegalArgumentException("work done id must not be initialized");
+        }
+        if (workDone.getWorkType() == null) {
+            throw new IllegalArgumentException("work type attribute must not be null");
+        }
         try (var connection = dataSource.getConnection();
-             var st1 = connection.prepareStatement(
+             var st = connection.prepareStatement(
                      "INSERT INTO WORK_TYPE (NAME, HOURLY_RATE, DESCRIPTION) VALUES (?, ?, ?)",
                      Statement.RETURN_GENERATED_KEYS)) {
 
-            st1.setString(1, workDone.getWorkType().getName());
-            st1.setDouble(2, workDone.getWorkType().getHourlyRate());
-            st1.setString(3, workDone.getWorkType().getDescription());
-            st1.executeUpdate();
-            var rs1 = st1.getGeneratedKeys();
-            rs1.next();
-            long workTypeID = rs1.getLong(1);
+            st.setString(1, workDone.getWorkType().getName());
+            st.setDouble(2, workDone.getWorkType().getHourlyRate());
+            st.setString(3, workDone.getWorkType().getDescription());
+            st.executeUpdate();
+            var rs = st.getGeneratedKeys();
+            rs.next();
+            Long workTypeID = rs.getLong(1);
             workDone.getWorkType().setId(workTypeID);
 
             insertWorkDone(workDone, connection, workTypeID);
@@ -40,24 +45,27 @@ public class WorkDao {
         }
     }
 
-    private void insertWorkDone(WorkDone workDone, Connection connection, long workTypeID) {
-        try (var st2 = connection.prepareStatement(
+    private void insertWorkDone(WorkDone workDone, Connection connection, Long workTypeID) {
+        try (var st = connection.prepareStatement(
                      "INSERT INTO WORK_DONE (WT_ID, WORK_START, WORK_END, DESCRIPTION) VALUES (?, ?, ?, ?)",
                 Statement.RETURN_GENERATED_KEYS)) {
-            st2.setLong(1, workTypeID);
-            st2.setTimestamp(2, Timestamp.valueOf(workDone.getWorkStart()));
-            st2.setTimestamp(3, Timestamp.valueOf(workDone.getWorkEnd()));
-            st2.setString(4, workDone.getDescription());
-            st2.executeUpdate();
-            var rs2 = st2.getGeneratedKeys();
-            rs2.next();
-            workDone.setId(rs2.getLong(1));
+            st.setLong(1, workTypeID);
+            st.setTimestamp(2, Timestamp.valueOf(workDone.getWorkStart()));
+            st.setTimestamp(3, Timestamp.valueOf(workDone.getWorkEnd()));
+            st.setString(4, workDone.getDescription());
+            st.executeUpdate();
+            var rs = st.getGeneratedKeys();
+            rs.next();
+            workDone.setId(rs.getLong(1));
         } catch (SQLException e) {
             throw new RuntimeException("Failed to store Work Done" + workDone, e);
         }
     }
 
     public void delete(WorkDone workDone) {
+        if (workDone.getId() == null) {
+            throw new IllegalArgumentException("Work done has null ID");
+        }
         try (var connection = dataSource.getConnection();
              var st = connection.prepareStatement(
                      "DELETE FROM WORK_DONE " +
@@ -65,28 +73,27 @@ public class WorkDao {
             st.setLong(1, workDone.getId());
             st.executeUpdate();
         } catch (SQLException ex) {
-            throw new RuntimeException("Failed to store employee " + workDone, ex);
+            throw new RuntimeException("Failed to delete work done " + workDone, ex);
         }
     }
 
-    private WorkType getWorkType(Connection connection, long workTypeID) {
-        try (var st2 = connection.prepareStatement(
+    private WorkType getWorkType(Connection connection, Long workTypeID) {
+        try (var st = connection.prepareStatement(
                 "SELECT ID, NAME, HOURLY_RATE, DESCRIPTION FROM WORK_TYPE WHERE ID = ?")) {
-            st2.setLong(1, workTypeID);
-            try (var rs2 = st2.executeQuery()) {
-                rs2.next();
+            st.setLong(1, workTypeID);
+            try (var rs = st.executeQuery()) {
+                rs.next();
                 return new WorkType(
                         workTypeID,
-                        rs2.getString("NAME"),
-                        rs2.getDouble("HOURLY_RATE"),
-                        rs2.getString("DESCRIPTION")
+                        rs.getString("NAME"),
+                        rs.getDouble("HOURLY_RATE"),
+                        rs.getString("DESCRIPTION")
                 );
             }
         } catch (Exception e1) {
             throw new RuntimeException();
         }
     }
-
 
     public List<WorkDone> findAll() {
         try (var connection = dataSource.getConnection();
@@ -113,7 +120,6 @@ public class WorkDao {
             throw new RuntimeException("Failed to load all works done", ex);
         }
     }
-
 
     private void initWorkDoneTable(TestDataGenerator testDataGenerator) {
         if (!workDoneTableExists("APP", "WORK_DONE")) {
@@ -172,6 +178,17 @@ public class WorkDao {
                     ")");
         } catch (SQLException ex) {
             throw new RuntimeException("Failed to create WORK_TYPE table", ex);
+        }
+    }
+
+    public void dropTable() {
+        try (Connection connection = dataSource.getConnection();
+             var st1 = connection.createStatement();
+             var st2 = connection.createStatement()) {
+            st1.executeUpdate("DROP TABLE APP.WORK_DONE");
+            st2.executeUpdate("DROP TABLE APP.WORK_TYPE");
+        } catch (SQLException e) {
+            throw new RuntimeException("failed to drop tables", e);
         }
     }
 }
